removed .env from .gitignore because I hate gatekeeping.
I'm also wondering as to why I did it in the first place. the point of open sourcing this repo was so that anyone can use this RAG system but gatekeep the main prompt template was as stupid as it can get. my bad, I take ownership of my foolery.